+++ src/plotly_toolbox/xy/__init__.py
@@ -1,4 +1,4 @@
-"""Cartesian (xy) plots: line, area, bar."""
+"""Cartesian (xy) plots: line, area, bar, scatter, histogram, distribution, heatmap."""
 
 from plotly_toolbox.xy.area import (
     AreaPlot,
@@ -16,6 +16,9 @@ from plotly_toolbox.xy.bars import (
     OneCategoryBars,
     TextPosition,
 )
+from plotly_toolbox.xy.distribution import BoxPlot, DistributionPlot, ViolinPlot
+from plotly_toolbox.xy.heatmap import Heatmap
+from plotly_toolbox.xy.histogram import Histogram
 from plotly_toolbox.xy.line import (
     CategoricalLines,
     LineMode,
@@ -24,6 +27,7 @@ from plotly_toolbox.xy.line import (
     MultiLinePlot,
     OneLinePlot,
 )
+from plotly_toolbox.xy.scatter import ScatterPlot
 
 __all__ = [
     "LineMode",
@@ -44,4 +48,10 @@ __all__ = [
     "OneCategoryBars",
     "ComparisonBars",
     "ComparasionBars",
+    "ScatterPlot",
+    "Histogram",
+    "DistributionPlot",
+    "BoxPlot",
+    "ViolinPlot",
+    "Heatmap",
 ]